Fixed unused_gate with as u64

+++ plonky2/src/gates/gate.rs
@@ -400,7 +400,7 @@ fn compute_filter<K: Field>(
     debug_assert!(group_range.contains(&row));
     group_range
         .filter(|&i| i != row)
-        .chain(many_selector.then_some(UNUSED_SELECTOR % characteristic as usize))
+        .chain(many_selector.then_some((UNUSED_SELECTOR as u64 % characteristic) as usize))
         .map(|i| K::from_canonical_usize(i) - s)
         .product()
 }
@@ -421,7 +421,7 @@ where
     debug_assert!(group_range.contains(&row));
     let v = group_range
         .filter(|&i| i != row)
-        .chain(many_selectors.then_some(UNUSED_SELECTOR % F::ORDER_U64 as usize))
+        .chain(many_selectors.then_some((UNUSED_SELECTOR as u64 % F::ORDER_U64) as usize))
         .map(|i| {
             let c = builder
                 .constant_extension(<F::Extension as AbstractField>::from_canonical_usize(i));
